Major rewrite of the widget logic. A candy.Widget is not clutter.Actor anymore but _has_ a clutter.Actor as child now. This results in making Widgets thread-safe and it removed the kaa.candy.init call. All set and get functions are replaced by properties.
The animation support is kind of broken right now. Widget.animate is gone
and there is no animation support in the xml definition. The animation
from clutter code is rewritten in Python but not fully integrated. At
this point it is unclear if Widget.animate will even come back.

A new backend submodule wraps all clutter related imports and is used
in cases where subclassing from clutter classes is needed.

Documentation is wrong in some cases and needs some update.


git-svn-id: svn://svn.freevo.org/kaa/trunk/candy@3394 a8f5125c-1e01-0410-8897-facf34644b8e