Second version of the gstreamer plugin (broken)
This widget uses gstreamer and the clutter sink directly. This
enables DVD playback and better control of the pipeline. But
there is a memory leak somehow, this code crashes: if we loose
the sink at creation gstreamer does not has a ref on its own and it
crashes. If we just keep the ref until the widget is destroyed,
it dies at that point (or later when we repeat this). Even
storing the sink forever crashes gstreamer with gobject messages
at some point. Maybe I'm doing something wrong here, but I don't
think so.